Refactor: organize business modules
